+++ spring-cloud-starter-tencent-polaris-router/src/test/java/com/tencent/cloud/polaris/router/config/properties/PolarisRuleBasedRouterPropertiesTest.java
@@ -18,8 +18,8 @@
 
 package com.tencent.cloud.polaris.router.config.properties;
 
-import org.junit.Before;
-import org.junit.Test;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -30,8 +30,8 @@ public class PolarisRuleBasedRouterPropertiesTest {
 
 	PolarisRuleBasedRouterProperties properties;
 
-	@Before
-	public void setUp() {
+	@BeforeEach
+	void setUp() {
 		properties = new PolarisRuleBasedRouterProperties();
 	}