+++ spring-cloud-tencent-rpc-enhancement/src/test/java/com/tencent/cloud/rpc/enhancement/feign/plugin/reporter/ExceptionPolarisReporterTest.java
@@ -22,28 +22,32 @@ import com.tencent.cloud.rpc.enhancement.feign.plugin.EnhancedFeignContext;
 import com.tencent.cloud.rpc.enhancement.feign.plugin.EnhancedFeignPluginType;
 import com.tencent.polaris.api.core.ConsumerAPI;
 import com.tencent.polaris.api.pojo.RetStatus;
+import com.tencent.polaris.api.rpc.ServiceCallResult;
 import feign.Request;
 import feign.Response;
-import org.junit.AfterClass;
-import org.junit.BeforeClass;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.MockedStatic;
 import org.mockito.Mockito;
-import org.mockito.junit.MockitoJUnitRunner;
+import org.mockito.junit.jupiter.MockitoExtension;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.doReturn;
 import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
 
 /**
  * Test for {@link ExceptionPolarisReporter}.
  *
  * @author Haotian Zhang
  */
-@RunWith(MockitoJUnitRunner.class)
+@ExtendWith(MockitoExtension.class)
 public class ExceptionPolarisReporterTest {
 
 	private static MockedStatic<ReporterUtils> mockedReporterUtils;
@@ -54,15 +58,15 @@ public class ExceptionPolarisReporterTest {
 	@InjectMocks
 	private ExceptionPolarisReporter exceptionPolarisReporter;
 
-	@BeforeClass
-	public static void beforeClass() {
+	@BeforeAll
+	static void beforeAll() {
 		mockedReporterUtils = Mockito.mockStatic(ReporterUtils.class);
 		mockedReporterUtils.when(() -> ReporterUtils.createServiceCallResult(any(Request.class), any(RetStatus.class)))
-				.thenReturn(null);
+				.thenReturn(mock(ServiceCallResult.class));
 	}
 
-	@AfterClass
-	public static void afterClass() {
+	@AfterAll
+	static void afterAll() {
 		mockedReporterUtils.close();
 	}
 
@@ -83,10 +87,16 @@ public class ExceptionPolarisReporterTest {
 		// mock response
 		Response response = mock(Response.class);
 
-		EnhancedFeignContext context = new EnhancedFeignContext();
-		context.setRequest(request);
-		context.setResponse(response);
+		EnhancedFeignContext context = mock(EnhancedFeignContext.class);
+		doReturn(request).when(context).getRequest();
+		doReturn(response).when(context).getResponse();
+		// test not report
 		exceptionPolarisReporter.run(context);
+		verify(context, times(0)).getRequest();
+		// test do report
+		doReturn(true).when(reporterProperties).isEnabled();
+		exceptionPolarisReporter.run(context);
+		verify(context, times(1)).getRequest();
 	}
 
 	@Test