+++ spring-cloud-tencent-rpc-enhancement/src/test/java/com/tencent/cloud/rpc/enhancement/resttemplate/EnhancedRestTemplateReporterTest.java
@@ -28,24 +28,28 @@ import com.tencent.cloud.common.metadata.MetadataContextHolder;
 import com.tencent.cloud.common.util.ApplicationContextAwareUtils;
 import com.tencent.cloud.rpc.enhancement.config.RpcEnhancementReporterProperties;
 import com.tencent.polaris.api.core.ConsumerAPI;
-import org.junit.AfterClass;
-import org.junit.Assert;
-import org.junit.Before;
-import org.junit.BeforeClass;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.checkerframework.checker.nullness.qual.NonNull;
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.MockedStatic;
 import org.mockito.Mockito;
-import org.mockito.junit.MockitoJUnitRunner;
+import org.mockito.junit.jupiter.MockitoExtension;
 
+import org.springframework.context.ApplicationContext;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpMethod;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.client.AbstractClientHttpResponse;
+import org.springframework.http.client.ClientHttpResponse;
+import org.springframework.web.client.DefaultResponseErrorHandler;
 import org.springframework.web.client.ResponseErrorHandler;
 
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.mock;
@@ -58,7 +62,7 @@ import static org.mockito.Mockito.when;
  *
  * @author lepdou 2022-09-06
  */
-@RunWith(MockitoJUnitRunner.class)
+@ExtendWith(MockitoExtension.class)
 public class EnhancedRestTemplateReporterTest {
 
 	private static MockedStatic<MetadataContextHolder> mockedMetadataContextHolder;
@@ -72,8 +76,11 @@ public class EnhancedRestTemplateReporterTest {
 	@InjectMocks
 	private EnhancedRestTemplateReporter enhancedRestTemplateReporter;
 
-	@BeforeClass
-	public static void beforeClass() {
+	@InjectMocks
+	private EnhancedRestTemplateReporter enhancedRestTemplateReporter2;
+
+	@BeforeAll
+	static void beforeAll() {
 		mockedApplicationContextAwareUtils = Mockito.mockStatic(ApplicationContextAwareUtils.class);
 		mockedApplicationContextAwareUtils.when(() -> ApplicationContextAwareUtils.getProperties(anyString()))
 				.thenReturn("caller");
@@ -89,15 +96,33 @@ public class EnhancedRestTemplateReporterTest {
 		mockedMetadataContextHolder.when(MetadataContextHolder::get).thenReturn(metadataContext);
 	}
 
-	@AfterClass
-	public static void afterClass() {
+	@AfterAll
+	static void afterAll() {
 		mockedApplicationContextAwareUtils.close();
 		mockedMetadataContextHolder.close();
 	}
 
-	@Before
-	public void before() {
+	@BeforeEach
+	void setUp() {
 		enhancedRestTemplateReporter.setDelegateHandler(delegate);
+	}
+
+	@Test
+	public void testSetApplicationContext() {
+		ApplicationContext applicationContext = mock(ApplicationContext.class);
+
+		// test no ResponseErrorHandler
+		when(applicationContext.getBeanNamesForType(any(Class.class)))
+				.thenReturn(new String[] {"enhancedRestTemplateReporter"});
+		enhancedRestTemplateReporter2.setApplicationContext(applicationContext);
+		assertThat(enhancedRestTemplateReporter2.getDelegateHandler()).isInstanceOf(DefaultResponseErrorHandler.class);
+
+		// test one other ResponseErrorHandler
+		when(applicationContext.getBeanNamesForType(any(Class.class)))
+				.thenReturn(new String[] {"enhancedRestTemplateReporter", "mockedResponseErrorHandler"});
+		when(applicationContext.getBean(anyString())).thenReturn(mock(MockedResponseErrorHandler.class));
+		enhancedRestTemplateReporter2.setApplicationContext(applicationContext);
+		assertThat(enhancedRestTemplateReporter2.getDelegateHandler()).isInstanceOf(MockedResponseErrorHandler.class);
 	}
 
 	@Test
@@ -105,10 +130,10 @@ public class EnhancedRestTemplateReporterTest {
 		when(delegate.hasError(any())).thenReturn(true);
 
 		MockedClientHttpResponse response = new MockedClientHttpResponse();
-		Assert.assertTrue(enhancedRestTemplateReporter.hasError(response));
+		assertThat(enhancedRestTemplateReporter.hasError(response)).isTrue();
 
 		String realHasError = response.getHeaders().getFirst(EnhancedRestTemplateReporter.HEADER_HAS_ERROR);
-		Assert.assertEquals("true", realHasError);
+		assertThat(realHasError).isEqualTo("true");
 	}
 
 	@Test
@@ -156,21 +181,21 @@ public class EnhancedRestTemplateReporterTest {
 		verify(delegate).handleError(uri, HttpMethod.GET, response);
 	}
 
-	class MockedClientHttpResponse extends AbstractClientHttpResponse {
+	static class MockedClientHttpResponse extends AbstractClientHttpResponse {
 
-		private HttpHeaders headers;
+		private final HttpHeaders headers;
 
 		MockedClientHttpResponse() {
 			this.headers = new HttpHeaders();
 		}
 
 		@Override
-		public int getRawStatusCode() throws IOException {
+		public int getRawStatusCode() {
 			return 0;
 		}
 
 		@Override
-		public String getStatusText() throws IOException {
+		public String getStatusText() {
 			return null;
 		}
 
@@ -193,5 +218,13 @@ public class EnhancedRestTemplateReporterTest {
 		public HttpStatus getStatusCode() throws IOException {
 			return HttpStatus.OK;
 		}
+	}
+
+	private static class MockedResponseErrorHandler extends DefaultResponseErrorHandler {
+
+		@Override
+		public void handleError(@NonNull ClientHttpResponse response) {
+		}
+
 	}
 }