+++ CHANGELOG.md
@@ -4,3 +4,7 @@
 - [docs:Add code coverage blocking.](https://github.com/Tencent/spring-cloud-tencent/pull/837)
 - [docs:Add license checker GitHub Action.](https://github.com/Tencent/spring-cloud-tencent/pull/841)
 - [refactor:move loadbalancer to discovery module.](https://github.com/Tencent/spring-cloud-tencent/pull/845)
+- [feature:add PolarisRateLimiterLimitedFallback spi. ](https://github.com/Tencent/spring-cloud-tencent/pull/858)
+- [fix:fix the error capture of rate limit exception.](https://github.com/Tencent/spring-cloud-tencent/pull/861)
+- [feat:enable stat reporting as default.](https://github.com/Tencent/spring-cloud-tencent/pull/864)
+- [refactor:update to junit 5.](https://github.com/Tencent/spring-cloud-tencent/pull/867)